Move log and config objects
