+++ src/environments/environment.ts
@@ -4,7 +4,7 @@
  * `ng build --env=prod` then `environment.prod.ts` will be used instead.
  * The list of which env maps to which file can be found in `.angular-cli.json`.
  */
-import {LogLevel} from '../app/models/log-level';
+import {LogLevel} from '../app/core/errors/log-level';
 
 export const environment = {
   production: false,
